+++ sdk/azcore/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History
 
+## 1.11.2 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
 ## 1.11.1 (2024-04-02)
 
 ### Bugs Fixed